extract tui logic into tui package
